Move GASActionBufferData to separate file
